+++ src/main/java/com/huongnguyen/repository/UserRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByEmail(String email);
 
@@ -25,5 +25,5 @@ public interface UserRepository extends JpaRepository<User, Integer> {
     List<UserResponse> findAllUsersWithRoles();
 
 
-    Optional<User> findById(Integer id);
+    Optional<User> findById(Long id);
 }
