+++ src/main/java/com/huongnguyen/service/impl/UserServiceImpl.java
@@ -41,7 +41,7 @@ public class UserServiceImpl implements UserService {
     }
 
 
-    public User getUserById(Integer id){
+    public User getUserById(Long id){
         return userRepository.findById(id)
                 .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_FOUND));
     }
